30: Add useFetchCategories hook
